LinkedIn-Auto-Applicant/requirements.txt
# LinkedIn Auto-Applicant Requirements
flask==3.0.0
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
pandas==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
flask-wtf==1.2.1
werkzeug==3.0.1
gunicorn==21.2.0
